+++ orchestrator/services/tasks.py
@@ -136,7 +136,7 @@ class CeleryJobWorkerStatus(WorkerStatus):
         self.number_of_workers_online = len(stats)
 
         def sum_items(d: dict) -> int:
-            return sum(len(l) for _, l in d.items())
+            return sum(len(l) for _, l in d.items()) if d else 0
 
         self.number_of_queued_jobs = sum_items(inspection.scheduled()) + sum_items(inspection.reserved())
         self.number_of_running_jobs = sum(len(tasks) for w, tasks in inspection.active().items())