+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+
+- Allow user to extend the default translation set, instead of overwriting everything  
+
 ## [0.4.0-rc1](https://github.com/workfloworchestrator/orchestrator-core/tree/0.4.0-rc1) (2022-03-08)
 
 [Full Changelog](https://github.com/workfloworchestrator/orchestrator-core/compare/0.3.8...0.4.0-rc1)